sem 1/hw3/src/main/src/main/java/com/spbau/wimag/hw3/SecondPartTasks.java
package com.spbau.wimag.hw3;

import java.io.IOException;
import java.nio.file.Files;
import java.nio.file.Paths;
import java.util.*;
import java.util.regex.Matcher;
import java.util.stream.Collectors;
import java.util.stream.DoubleStream;
import java.util.stream.IntStream;
import java.util.stream.Stream;

public final class SecondPartTasks {

    private SecondPartTasks() {
    }

    // Найти строки из переданных файлов, в которых встречается указанная подстрока.
    public static List<String> findQuotes(List<String> paths, CharSequence sequence) {
        return paths.stream().flatMap(s -> {
            Stream<String> res = null;
            try {
                res = Files.lines(Paths.get(s));
            } catch (IOException e) {
                e.printStackTrace();
            }
            return res;
        }).filter(s -> s.contains(sequence)).collect(Collectors.toList());
    }

    // В квадрат с длиной стороны 1 вписана мишень.
    // Стрелок атакует мишень и каждый раз попадает в произвольную точку квадрата.
    // Надо промоделировать этот процесс с помощью класса java.util.Random и посчитать, какова вероятность попасть в мишень.
    public static double piDividedBy4() {
        int attempts = 100000;
        Random random = new Random();
        return (double) random.doubles(attempts).filter(x -> {
            double dx = Math.abs(x - 0.5);
            double dy = Math.abs(random.nextDouble() - 0.5);

            return (dx * dx + dy * dy) <= 0.25;
        }).count() / (double) attempts;
    }

    // Дано отображение из имени автора в список с содержанием его произведений.
    // Надо вычислить, чья общая длина произведений наибольшая.
    public static String findPrinter(Map<String, List<String>> compositions) {
        return compositions.entrySet().stream().max(Comparator.comparingInt(e -> e.getValue().stream()
                .mapToInt(String::length).sum())).get().getKey();
    }

    // Вы крупный поставщик продуктов. Каждая торговая сеть делает вам заказ в виде Map<Товар, Количество>.
    // Необходимо вычислить, какой товар и в каком количестве надо поставить.
    public static Map<String, Integer> calculateGlobalOrder(List<Map<String, Integer>> orders) {
        return orders.stream().flatMap(x -> x.entrySet().stream()).collect(Collectors
                .toMap(Map.Entry::getKey, Map.Entry::getValue, (a, b) -> a + b));
    }
}
